cosmetics, make participantid a long plus avoid race condition...

+++ OpenSim/Addons/os-webrtc-janus/Janus/JanusViewerSession.cs
@@ -54,8 +54,8 @@ namespace osWebRtcVoice
         public OMV.UUID AgentId { get; set; }
 
         // Janus keeps track of the user by this ID
-        public int ParticipantId { get; set; }
-//        public long ParticipantId { get; set; }
+//        public int ParticipantId { get; set; }
+        public long ParticipantId { get; set; }
 
         // Connections to the Janus server
         public JanusSession Session { get; set; }
